pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Randi Guarcas <rguarcas18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.2"
flask-restful = "^0.3.9"
requests = "^2.30.0"
python-slugify = "^8.0.1"
flask-cors = "^3.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
